requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.1.0
bleach==4.1.0
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
entrypoints==0.4
executing==0.8.2
folium==0.12.1.post1
fonttools==4.29.1
idna==3.3
ipykernel==6.9.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
munch==2.5.0
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.7.1
notebook==6.4.8
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.0
prometheus-client==0.13.1
prompt-toolkit==3.0.27
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyproj==3.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
Send2Trash==1.8.0
Shapely~=1.8.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.4
terminado==0.13.1
testpath==0.5.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
tensorflow~=2.7.0
pip~=21.2.3
wheel~=0.37.1
rsa~=4.8
pyasn1~=0.4.8
gast~=0.5.3
h5py~=3.6.0
keras~=2.7.0
wrapt~=1.14.0
setuptools~=57.4.0
cachetools~=5.0.0
Markdown~=3.3.6
oauthlib~=3.2.0
Werkzeug~=2.0.3
tensorboard~=2.8.0
flatbuffers~=2.0
astunparse~=1.6.3
termcolor~=1.1.0
zipp~=3.7.0
gym~=0.21.0
pygame
osmnx~=1.1.2
fiona~=1.8.21
geopandas~=0.10.2
scikit-learn~=1.0.2
haversine~=2.5.1